refactor(app): extract headless-shared from headless.ts
Moves shared infra (HeadlessDeps, executor wiring, restore/preempt/flag-parsing/tracking helpers, ANSI constants) out of headless.ts into headless-shared.ts, behavior unchanged.
One atomic slice of the step-4 large-file decomposition.
